requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.23
sqlalchemy>=2.0
pandas>=2.2
openpyxl>=3.1
python-multipart>=0.0.9
requests>=2.31
python-dotenv>=1.0
